// src/app/api/models/advanced-low-stock-product-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AdvancedLowStockProductDto {
  averageConsumptionRate: number;
  currentStock: number;
  daysUntilStockout: number;
  lastMovementDate: string;
  minStockLevel: number;
  productId: number;
  productName: string;
  reorderPoint: number;
  suggestedReorderQuantity: number;
}