// src/app/api/models/top-movers-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TopMoversDto {
  adjustmentMovements: number;
  inMovements: number;
  movementValue: number;
  outMovements: number;
  productId: number;
  productName: string;
  totalMovements: number;
  trend: 'up' | 'down' | 'stable';
  turnoverRate: number;
  usageMovements: number;
}